- update database schema

+++ Training.DataAccess/Entities/OrderDetail.cs
@@ -12,9 +12,9 @@ namespace Training.DataAccess.Entities
     [Table("OrderDetails")]
     public class OrderDetail : IBaseEntity
     {
-        public int Id { get; set; }
-        public int OrderId { get; set; }
-        public int ProductId { get; set; }
+        public long Id { get; set; }
+        public long OrderId { get; set; }
+        public long ProductId { get; set; }
         public float UnitPrice { get; set; }
         public int Quantity { get; set; }
         public Order Order { get; set; }
